+++ resources/views/auth/passwords/reset.blade.php
@@ -3,7 +3,11 @@
 
 @section('content')
     <p class="m-t-lg text-black">{{ __('Reset Password') }}</p>
-    <form class="m-t-md" role="form" method="POST"
+    <form class="m-t-md"
+          role="form"
+          method="POST"
+          data-controller="layouts--form"
+          data-action="layouts--form#submit"
           action="{{ route('platform.password.email') }}">
         @csrf
         <div class="form-group {{ $errors->has('email') ? ' is-invalid' : '' }}">