+++ resources/views/container/layouts/base.blade.php
@@ -12,12 +12,16 @@
     </ul>
 @stop
 @section('content')
-    <section>
-        <form id="post-form" method="post" enctype="multipart/form-data">
+        <form id="post-form"
+              method="post"
+              enctype="multipart/form-data"
+              data-controller="layouts--form"
+              data-action="layouts--form#submit"
+              data-text-validation="{{__('Please check the entered data, it may be necessary to specify in other languages.')}}"
+        >
             {!! $screen->build() !!}
             @csrf
         </form>
-    </section>
     <div id="modals-container">
         @stack('modals-container')
     </div>